--- src/main/java/com/example/dp/domain/order/dto/request/OrderRequestDto.java
@@ -1,11 +1,0 @@
-package com.example.dp.domain.order.dto.request;
-
-public class OrderRequestDto {
-
-    private String name; // 메뉴 이름
-
-    private Integer menuCounts; //해당 메뉴의 갯수
-
-    private Integer priceSum; //총 가격
-
-}
