+++ src/main/java/com/example/dp/domain/order/service/OrderService.java
@@ -1,8 +1,7 @@
 package com.example.dp.domain.order.service;
 
 
-import com.example.dp.domain.order.dto.OrderResponseDto;
-import com.example.dp.domain.order.dto.request.OrderRequestDto;
+import com.example.dp.domain.order.dto.response.OrderResponseDto;
 import com.example.dp.domain.order.dto.response.OrderSimpleResponseDto;
 import com.example.dp.domain.user.entity.User;
 import java.util.List;
